Dynamic response audio to text working
- nagbabago na yung loading once audio to text
- nilipat most ang mga ffiles sa utilities folder
-  deleted some files on mysqli/sql-table-queries
- error handling of text-text now in  error-handling.php
- process of text-text now in trnaslation.php
- error handling of user manipulating inspect element on choices now added
- basically nilinis ko lang talaga yung codes, like require,  heirarchy of functiions,  etcs
