+++ history_audio.php
@@ -1,114 +1,23 @@
 <?php require("mysql/mysqli_session.php"); 
-$current_page = basename($_SERVER['PHP_SELF']);
-?>
-<?php if (!isset($_SESSION['username'])) {
-    
-  header("location: index.php");
-  exit(); 
-}?>
+    $current_page = basename($_SERVER['PHP_SELF']);
+    if (!isset($_SESSION['username'])) {
+        header("location: index.php");
+        exit(); 
+    }
 
-<?php
+    function dd($item){
+        var_dump($item);
+        exit();
+    }
 
-function dd($item){
-    var_dump($item);
-    exit();
-}
+require("utilities/Translator_Functions.php");
 
-require "Translator_Functions.php";
-
-require "languages.php";
-
-  $id = is_array($_SESSION['user_id']) ? $_SESSION['user_id']['user_id'] : $_SESSION['user_id'];
+$id = is_array($_SESSION['user_id']) ? $_SESSION['user_id']['user_id'] : $_SESSION['user_id'];
 
 
 // Translation history for audio to text 
 $history = mysqli_query($dbcon, "SELECT * FROM text_translations t INNER JOIN audio_files a ON t.file_id = a.file_id WHERE t.user_id = $id AND a.user_id = $id AND t.from_audio_file = 1 ORDER BY translation_date DESC");
 
-// Language Translation, please check https://rapidapi.com/dickyagustin/api/text-translator2 for more information.
-// Translate the input
-
-/*
-if($_SERVER["REQUEST_METHOD"] == "POST"){
-
-    // required for uploading the file
-    $path=$_FILES['user_file']['name']; // file
-    $pathsize = $_FILES['user_file']['size']; // file size
-    $userid = $_SESSION['user_id']; // user id needed to separate all files between each user by appending userid to filename
-    
-
-	$src_lang = ($_POST['src'] == 'auto') ? "auto" : $audio_src_lang_codes[$_POST["src"]];
-	$trg_lang = $audio_trgt_lang_codes[$_POST["target"]] ?? ''; 
-
-
-    // Checks whether checkbox is checked or not
-    $removeBGM = ISSET($_POST["removeBGM"]) ?  "on" : "off";
-    
-
-	
-	// step 1: loading
-    // error handlings first before proceeding to the main process
-        // will automatically halt the process once error caught
-	ErrorHandling::checkLanguageChosen();
-	ErrorHandling::validateFormat($path);
-	ErrorHandling::checkFolder();
-
-
-    Translator::db_insertAudioFile($path, $userid, $pathsize);
-    $newFile = Translator::createNewFilename($path, $userid);
-
-    // step 2: extracting vocals
-    # Extract vocals if checkbox is checked
-    if ($removeBGM === "on") {
-        Translator::getVocals($newFile);
-    }
-
-    // step 3: transcribing audio
-    //  $transcript contains the text and language
-    $transcript = Translator::uploadAndTranscribe($newFile, $removeBGM, $src_lang, $_POST['modelSize']);
-
-
-
-    // step 4: translating text
-    $result = Translator::translate($transcript['text'], $transcript['language'], $trg_lang);
-
-
-
-    //  STEP 6
-    // after translating, find the language name in the common language array
-    $key = array_search($transcript['language'], array_column($common_languages, 'code'));    
-    if ($_POST['src'] == 'auto') {
-        $source_lang = $common_languages[$key]['name']; // extract the lang name from array if user chooses auto-detect
-    } else {
-        $source_lang = $_POST['src']; 
-    }
-
-    $target_lang = $_POST['target'];
-
-    $isFromAudio = TRUE;
-    
-    $get_fileid = "SELECT file_id FROM audio_files WHERE user_id = '$id' ORDER BY file_id DESC LIMIT 1";
-    $fileresult = mysqli_query($dbcon, $get_fileid);
-
-    $row = mysqli_fetch_assoc($fileresult);
-
-    $query_insert1 = mysqli_prepare($dbcon, "INSERT INTO text_translations(file_id, user_id, from_audio_file, original_language, translated_language,
-    translate_from, translate_to, translation_date) VALUES (?, ?, ?, ?, ?, ?, ?, NOW())");
-    
-    mysqli_stmt_bind_param($query_insert1, 'iiissss', $row['file_id'], $id, $isFromAudio, $source_lang, $target_lang, $transcript['text'], $result);
-    mysqli_stmt_execute($query_insert1);
-
-	$response = array(
-		"validatingInputs" => true, // Replace with actual validation status
-		"uploadingFile" => true, // Replace with actual upload status
-		"extractingVocals" => true, // Replace with actual extraction status
-		"transcribingAudio" => true, // Replace with actual transcription status
-		"translatingText" => true // Replace with actual translation status
-	);
-	
-    logs("audio-to-text", $_SESSION['username'], $dbcon);
-    header("Location: history_audio.php?translated=1");
-}
-*/
 
 ?>
 
@@ -183,7 +92,8 @@ if($_SERVER["REQUEST_METHOD"] == "POST"){
                                 case 5: // it's what the programmers say, "WHY IS IT NOT WORKING??"
                                     echo "Audio File not processed well. Please try again.";
                                     break;
-                                default;
+                                case 6: // user added unprovided choices
+                                    echo "Please choose only on the provided models/languages.";
                                     break;
                             }
                         }
@@ -199,7 +109,7 @@ if($_SERVER["REQUEST_METHOD"] == "POST"){
                         <p id="loadingModalMessage">Loading....</p>
                     </div>
                 </div>
-            <form enctype="multipart/form-data" id="form" method = "POST" onsubmit="showLoading()">
+            <form enctype="multipart/form-data" id="form" action="utilities/audio_translation.php" method = "POST" onsubmit="showLoading()">
 
             <label>  
 			Model Size:
@@ -329,7 +239,6 @@ if($_SERVER["REQUEST_METHOD"] == "POST"){
                                 
     <script src="scripts/index.js"></script>
     <script src="scripts/delete.js"></script>
-    <!--<script src="scripts/translation_process.js"></script>-->
     <script src="scripts/translation_process.js"></script>
 
 </body>
