+++ mysql/sql-table-queries/AngAmingDataBase - (2) insert table.sql
