+++ mysql/sql-table-queries/AngAmingDatabase - (3) alter_datatype.sql
