--- mysql/sql-table-queries/create table2.sql
@@ -1,1 +1,0 @@
-CREATE DATABASE AudioToText;
