+++ sidebar.php
@@ -20,7 +20,7 @@
         </ul>
         <ul class="side-menu">
             <li>
-                <a href="logout.php" class="logout">
+                <a href="utilities/logout.php" class="logout">
                     <i class='bx bx-log-out-circle'></i>
                     Logout
                 </a>
